+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.package_name}}/__init__.py
@@ -1,1 +1,1 @@
-from .meta import __version__
+from ._meta import __version__
