+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.package_name}}/_meta.py
