+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.package_name}}/api/module_setup.py
@@ -2,7 +2,7 @@
     api app module
 """
 from fastapi import APIRouter, FastAPI
-from ..meta import api_vtag
+from .._meta import api_vtag
 
 
 def setup_api(app: FastAPI):
